app/requirements.txt
fastapi
uvicorn[standard]
prometheus-client
psycopg2-binary